+++ app/lib/main.dart
@@ -21,7 +21,7 @@ import 'package:spendwise/provider/token_provider.dart';
 import 'package:spendwise/screens/login_screen.dart';
 import 'package:spendwise/screens/main_screen.dart';
 import 'package:spendwise/screens/register_screen_screen.dart';
-import 'package:spendwise/screens/edit_data_screens/select_%20monetary_unit.dart';
+import 'package:spendwise/screens/edit_screens/select_%20monetary_unit.dart';
 import 'package:spendwise/screens/static_screens/startup_screen.dart';
 
 // utils
