+++ app/lib/screens/pages/account_page.dart
@@ -10,8 +10,8 @@ import 'package:spendwise/provider/transaction_provider.dart';
 import 'package:spendwise/provider/user_provider.dart';
 
 // Screens & Widgets
-import 'package:spendwise/screens/edit_data_screens/add_update_account_screen.dart';
-import 'package:spendwise/screens/edit_data_screens/update_account_screen.dart';
+import 'package:spendwise/screens/edit_screens/edit_account/edit_acccount_screen.dart';
+import 'package:spendwise/screens/edit_screens/update_account_screen.dart';
 import 'package:spendwise/screens/view_all_transaction_screen.dart';
 import 'package:spendwise/widgits/cards/account_card.dart';
