+++ app/lib/screens/static_screens/setting_screen.dart
@@ -17,7 +17,7 @@ import 'package:spendwise/provider/user_provider.dart';
 
 // Screens & Widgets
 import 'package:spendwise/screens/login_screen.dart';
-import 'package:spendwise/screens/edit_data_screens/select_%20monetary_unit.dart';
+import 'package:spendwise/screens/edit_screens/select_%20monetary_unit.dart';
 import 'package:spendwise/widgits/cards/tile_card.dart';
 import 'package:spendwise/widgits/cards/update_card.dart';
