+++ app/lib/widgits/cards/balance_card.dart
@@ -13,7 +13,7 @@ import 'package:spendwise/provider/transaction_provider.dart';
 import 'package:spendwise/provider/user_provider.dart';
 
 // Screens & Widgets
-import 'package:spendwise/screens/edit_data_screens/add_update_transaction_screen.dart';
+import 'package:spendwise/screens/edit_screens/edit_transactions/edit_transaction_screen.dart';
 
 // ignore: must_be_immutable
 class BalanceCards extends ConsumerWidget {
